+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Unreleased:**
 
 * Support dictionary and tuple assignments in solution substitution in Otter Assign per [#587](https://github.com/ucbds-infra/otter-grader/issues/587)
+* Add the `pdf` argument to `ottr::export` in R assignments created with Otter Assign per [#440](https://github.com/ucbds-infra/otter-grader/issues/440)
 
 **v4.2.1:**